% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestPlot.R
\name{getTestPlot}
\alias{getTestPlot}
\title{Plot inference result from the VCCP model}
\usage{
getTestPlot(vccp_result)
}
\arguments{
\item{vccp_result}{A list generated from \code{\link{vccp.fun}}.}
}
\description{
\code{getTestPlot} plots possible change points in the functional connectivity
of a multivariate time series and their
significance generated by a VCCP model.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = vccp.fun(data, "NBS", test = "V")
getTestPlot(result)

result.2 = vccp.fun(data, "NBS", test = "B")
getTestPlot(result.2)
}
\seealso{
\code{\link{vccp.fun}}
}
