% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VuongTest.R
\name{GetTestPlot.Vuong}
\alias{GetTestPlot.Vuong}
\title{Plot inference result from VCCP(Vuong test)}
\usage{
GetTestPlot.Vuong(test_result, T, sig_alpha = 0.05)
}
\arguments{
\item{test_result}{A 4-column dataframe; usually coming from the result
of \code{\link{VCCP}} with \code{test} specified as 'V' or \code{\link{TestPoints.Vuong}}.}

\item{T}{An integer; the length of the single-subject time series data.}

\item{sig_alpha}{A decimal between 0 and 1; the significance level of the Vuong test, which
should be consistent with \code{sig_alpha} in \code{\link{VCCP}}.}
}
\description{
\code{GetTestPlot.Vuong} plots the possible change points in the connection network
and their significance tested by the Vuong method.
}
\details{
For each candidate, the Vuong test will perform two sub-tests based on the comparison
between the left VC (i.e., Vine copula distribution) and the whole VC (blue dots) and between the right VC and the whole VC
(green dots). If either of the tests suggests the whole VC performs better (i.e., the sign
of the likelihood ratio statistic is greater than 0) or the left/right VC is not significantly
better than the whole VC (i.e., the blue/green dot falls below the \code{log(1/sig_alpha)} line),
we would say the corresponding candidate fails the test and is not a significant change point.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = VC.NBS.FindPoints(data, 30)
inference = TestPoints.Vuong(result[[1]], data, delta=30)
GetTestPlot.Vuong(inference, 180)

inference.2 = VCCP(data, method = "OBS", 30, test = "V")
GetTestPlot.Vuong(inference.2, 180)

}
\seealso{
\code{\link{VCCP}}, \code{\link{TestPoints.Vuong}}
}
