% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootsTest.R
\name{GetTestPlot_Boot}
\alias{GetTestPlot_Boot}
\title{Plot inference result from VCCP(Stationary Bootstrap)}
\usage{
GetTestPlot_Boot(test_result, T)
}
\arguments{
\item{test_result}{A 5-column dataframe; usually coming from the result
of \code{\link{VCCP}} with \code{test} specified as 'B'.}

\item{T}{An integer; the length of the single-subject time series data.}
}
\description{
\code{GetTestPlot_Boot} plots the possible change points in the connection network
and their significance tested by the Stationary Bootstrap method.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = VC_NBS_FindPoints(data, 30)
inference = TestPoints_Boot(result[[1]], data, delta=30)
GetTestPlot_Boot(inference, 180)

inference.2 = VCCP(data, method = "OBS", 30, test = "B")
GetTestPlot_Boot(inference.2, 180)
}
\seealso{
\code{\link{VCCP}}, \code{\link{TestPoints_Boot}}
}
