% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VuongTest.R
\name{GetTestPlot_Vuong}
\alias{GetTestPlot_Vuong}
\title{Plot inference result from VCCP(Vuong test)}
\usage{
GetTestPlot_Vuong(test_result, T, sig_alpha = 0.05)
}
\arguments{
\item{test_result}{A 4-column dataframe; usually coming from the result
of \code{\link{VCCP}} with \code{test} specified as 'V' or \code{\link{TestPoints_Vuong}}.}

\item{T}{An integer; the length of the single-subject time series data.}

\item{sig_alpha}{A decimal between 0 and 1; the significance level of the Vuong test, which
should be consistent with \code{sig_alpha} in \code{\link{VCCP}}.}
}
\description{
\code{GetTestPlot_Boot} plots the possible change points in the connection network
and their significance tested by the Vuong method.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = VC_NBS_FindPoints(data, 30)
inference = TestPoints_Vuong(result[[1]], data, delta=30)
GetTestPlot_Vuong(inference, 180)

inference.2 = VCCP(data, method = "OBS", 30, test = "V")
GetTestPlot_Vuong(inference.2, 180)

}
\seealso{
\code{\link{VCCP}}, \code{\link{TestPoints_Vuong}}
}
