% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Data.R
\name{random.mvn.simulate.2.changes}
\alias{random.mvn.simulate.2.changes}
\title{Simulate MVN data with 2 change points}
\usage{
random.mvn.simulate.2.changes(nobs, n.ser, seed)
}
\arguments{
\item{nobs}{A positive integer number equal to the length of the time series data (T).
\code{nobs} must be multiple times of 3 since change points occur at T/3
and 2T/3.}

\item{n.ser}{A positive integer number indicating the dimensionality of the
time series. \code{n.ser} must be larger than 8 as \code{random.mvn.simulate.2.changes}
generates 3 different networks among 8 connected nodes. Other variables are generated
to be independent of the 8 nodes all the time.}

\item{seed}{A positive integer number that ensures reproducibility.}
}
\value{
A \code{nobs} times \code{n.ser} matrix with 2 change points
at t=\code{nobs/3+1} and t=\code{nobs*2/3+1} in the network structure.
}
\description{
\code{random.mvn.simulate.2.changes} returns a multivariate normal
data with 2 change points in the connection network. This function
aims to display how VCCP model detects change points in examples.
}
\examples{
mvn <- random.mvn.simulate.2.changes(180, 8, seed = 101)
}
