% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootsTest.R
\name{TestPoints_Boot}
\alias{TestPoints_Boot}
\title{Test change points in connection network by Stationary Bootstrap}
\usage{
TestPoints_Boot(
  v_t_point,
  X_raw,
  delta,
  CDR = "D",
  trunc_tree = NA,
  family_set = 1,
  pre_white = 0,
  ar_num = 1,
  p = 0.3,
  N = 100,
  sig_alpha = 0.05
)
}
\arguments{
\item{v_t_point}{A integer vector of change point candidates, normally
from the first element of the list generated by \code{\link{VC_NBS_FindPoints}}.}

\item{X_raw}{A matrix with at least three columns. The first one must
be integers indicating time. If multiple subjects are included,
you shoud stack their data vertically with the corresponding
timestamps shown in the first column. The other two or more columns are
the multi-dimensional time series data (n*T times p).}

\item{delta}{An integer indicating the least distance between two
change points. This parameter is vital for the New Binary Segmentation (NBS) method
and must be specified. Normally \code{delta >= 5*(dim(X_raw)[2]-1)} is recommended to ensure sufficient
data when fitting the VC model.}

\item{CDR}{An upper letter specifying the tree type of the vine model:
"\code{D}" = D-Vine (default),
"\code{C}" = C-Vine,
"\code{R}" = R-Vine.}

\item{trunc_tree}{An integer or NA; level of truncation for the vine model.
If NA is specified (default), the vine contains \code{dim(X_raw)[2]-2} levels of trees.}

\item{family_set}{An integer vector of pair-copula families to select from.
The vector has to include at least one pair-copula family that allows for
positive and one that allows for negative dependence.
If \code{familyset} = 1 (default), only Gauss copula is selected and VCCP can only
detect changes in the linear correlation network.
Coding of pair-copula families is the same as in \code{\link[VineCopula]{BiCop}}.}

\item{pre_white, ar_num}{Integers helping fit (method: yule-walker) an
autoregressive time series model to preprocess the raw data. If \code{pre_white}=0(default),
no ar model is fitted. If \code{pre_white}=1, ar_num is the maximum order of model to fit.}

\item{p}{A decimal between 0 and 1; controlling the sampling block size in
Stationary Boostrap method (\code{rgeom(T,p)}). The larger p (default=0.3) is, the fewer time
points each bootstrap sample contains.}

\item{N}{An integer; specifying the number of Stationary Bootstrap samples (default=100).}

\item{sig_alpha}{A decimal between 0 and 1; significance level of the Stationary Bootsrtap test.}
}
\value{
A dataframe. The first column contains possible change point candiates;
the second one corresponds to the reduced BIC values (left VC + right VC - all_VC);
the third and the fourth columns are the lower and upper bound of reduced BIC values calculated by the
Stationary Bootsrtap test; and the fifth one is the inference result.
}
\description{
\code{TestPoints_Boot} returns a dataframe containing the inference result
of significant change points in the connection network among the multi/single-subject
time series data. It can be used after detecting possible change points
by \code{\link{VC_NBS_FindPoints}}.
}
\details{
If you get the change point candidates from \code{\link{VC_NBS_FindPoints}},
parameters used in \code{TestPoints_Boot} are strongly recommended to be consistent
with the ones specified when detecting possible candidates (i.e., \code{delta},
\code{CDR}, \code{trunc_tree}, \code{family_set}, \code{pre_white} and \code{ar_num}).
You can also combine the detection (NBS) and inference process (Stationary Bootstrap or Vuong Test)
together by \code{\link{VC_NBS}}.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = VC_NBS_FindPoints(data, 30)
inference = TestPoints_Boot(result[[1]], data, delta=30)
}
\seealso{
\code{\link{VC_NBS_FindPoints}}, \code{\link{TestPoints_Vuong}}, \code{\link{VC_NBS}}
}
