% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBS_FindPoints.R
\name{VC.NBS.FindPoints}
\alias{VC.NBS.FindPoints}
\title{Find change point candidates in the functional connectivity by VCCP (NBS) model}
\usage{
VC.NBS.FindPoints(
  X_raw,
  delta,
  CDR = "D",
  trunc_tree = NA,
  family_set = 1,
  pre_white = 0,
  ar_num = 1
)
}
\arguments{
\item{X_raw}{A matrix with at least three columns. The first one must
be integers indicating time. If multiple subjects are included,
you shoud stack their data vertically with the corresponding
timestamps shown in the first column. The other two or more columns are
the multi-dimensional time series data.}

\item{delta}{An integer indicating the least distance between two
change points. This parameter is vital for the New Binary Segmentation (NBS) method
and must be specified. Normally \code{delta >= 5*(dim(X_raw)[2]-1)} is recommended to ensure sufficient
data when fitting the VC model.}

\item{CDR}{An upper letter specifying the tree type of the vine model:
"\code{D}" = D-Vine (default),
"\code{C}" = C-Vine,
"\code{R}" = R-Vine.}

\item{trunc_tree}{An integer or NA; level of truncation for the vine model.
If NA is specified (default), the vine contains \code{dim(X_raw)[2]-2} levels of trees.}

\item{family_set}{An integer vector of pair-copula families to select from.
The vector has to include at least one pair-copula family that allows for
positive and one that allows for negative dependence.
If \code{familyset} = 1 (default), only Gauss copula is selected and VCCP can only
detect changes in the linear correlation network.
Coding of pair-copula families is the same as in \code{\link[VineCopula]{BiCop}}.}

\item{pre_white, ar_num}{Integers helping fit (method: yule-walker) an
autoregressive time series model to preprocess the raw data. If \code{pre_white}=0(default),
no ar model is fitted. If \code{pre_white}=1, ar_num is the maximum order of model to fit.}
}
\value{
An integer vector of detected possible change points.
}
\description{
\code{VC.NBS.FindPoints} returns possible change points in the correlation network
among the time series data.
}
\details{
The time series data can come from one subject, or from multiple subjects.
The VC model are fitted using \code{\link[VineCopula]{RVineStructureSelect}} or
\code{\link[CDVine]{CDVineCopSelect}}.
}
\examples{
data = cbind(1:180, random.mvn.simulate.2.changes(180,8,seed=101))
result = VC.NBS.FindPoints(data, 30)
}
\seealso{
\code{\link{TestPoints.Boot}}, \code{\link{TestPoints.Vuong}}, \code{\link{VCCP}}
}
