% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestPlot.R
\name{getTestPlot}
\alias{getTestPlot}
\title{Plot result from the VCCP model}
\usage{
getTestPlot(vccp_result)
}
\arguments{
\item{vccp_result}{A list generated from \code{\link{vccp.fun}}.}
}
\description{
This function plots change points in the functional connectivity
network structure between multivariate time series detected by the VCCP model.
}
\examples{
## Simulate MVN data with 2 change points
data = cbind(1:180, mvn.sim.2.cps(180,8,seed=101))
## Change point detection using VCCP
result = vccp.fun(data, "NBS", test = "V")
## Plot the result
getTestPlot(result)

## Change point detection using NBS and stationary bootstrap for inference
result.2 = vccp.fun(data, "NBS", test = "B")
## Plot the result
getTestPlot(result.2)
}
\seealso{
\code{\link{vccp.fun}}
}
