% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Data.R
\name{random.mvn.simulate.2.changes}
\alias{random.mvn.simulate.2.changes}
\title{Simulate MVN data with 2 change points}
\usage{
random.mvn.simulate.2.changes(nobs, n.ser, seed)
}
\arguments{
\item{nobs}{An integer; equal to the length of the time series data (T).
\code{nobs} must be multiple times of 3 since change points occur at T/3
and 2T/3.}

\item{n.ser}{An integer indicating the number of variables in the
time series. \code{n.ser} must be larger than 8 as \code{random.mvn.simulate.2.changes}
generates 3 different networks on at most 8 connected nodes. If n.ser>8,
nodes on columns other than the first 8 are supposed to be independent from
other nodes all the time.}

\item{seed}{An integer; set for the purpose of reproducibility study.}
}
\value{
A \code{nobs} times \code{n.ser} matrix with 2 change points
at t=\code{nobs/3+1} and t=\code{nobs*2/3+1} in the network structure.
}
\description{
\code{random.mvn.simulate.2.changes} returns a multivariate normal
data with 2 change points in the connection network. This function
aims to display how VCCP model detects change points in examples.
}
\examples{
mvn <- random.mvn.simulate.2.changes(180, 8, seed = 101)
}
